Move some namespaces on Exceptions
