+++ src/Bundle/Controller/ResourceController.php
@@ -16,9 +16,9 @@ namespace Sylius\Bundle\ResourceBundle\Controller;
 use Doctrine\Persistence\ObjectManager;
 use FOS\RestBundle\View\View;
 use Sylius\Bundle\ResourceBundle\Event\ResourceControllerEvent;
-use Sylius\Component\Resource\Exception\DeleteHandlingException;
-use Sylius\Component\Resource\Exception\UpdateHandlingException;
 use Sylius\Resource\Doctrine\Persistence\RepositoryInterface;
+use Sylius\Resource\Exception\DeleteHandlingException;
+use Sylius\Resource\Exception\UpdateHandlingException;
 use Sylius\Resource\Factory\FactoryInterface;
 use Sylius\Resource\Metadata\MetadataInterface;
 use Sylius\Resource\Model\ResourceInterface;
