+++ src/Component/legacy/spec/Exception/RaceConditionExceptionSpec.php
@@ -14,7 +14,7 @@ declare(strict_types=1);
 namespace spec\Sylius\Component\Resource\Exception;
 
 use PhpSpec\ObjectBehavior;
-use Sylius\Component\Resource\Exception\UpdateHandlingException;
+use Sylius\Resource\Exception\UpdateHandlingException;
 
 final class RaceConditionExceptionSpec extends ObjectBehavior
 {
