+++ src/Component/legacy/src/Exception/UpdateHandlingException.php
@@ -13,32 +13,10 @@ declare(strict_types=1);
 
 namespace Sylius\Component\Resource\Exception;
 
-class UpdateHandlingException extends \Exception
-{
-    protected string $flash;
+class_exists(\Sylius\Resource\Exception\UpdateHandlingException::class);
 
-    protected int $apiResponseCode;
-
-    public function __construct(
-        string $message = 'Ups, something went wrong during updating a resource, please try again.',
-        string $flash = 'something_went_wrong_error',
-        int $apiResponseCode = 400,
-        int $code = 0,
-        ?\Exception $previous = null,
-    ) {
-        parent::__construct($message, $code, $previous);
-
-        $this->flash = $flash;
-        $this->apiResponseCode = $apiResponseCode;
-    }
-
-    public function getFlash(): string
+if (false) {
+    class UpdateHandlingException extends \Sylius\Resource\Exception\UpdateHandlingException
     {
-        return $this->flash;
-    }
-
-    public function getApiResponseCode(): int
-    {
-        return $this->apiResponseCode;
     }
 }
