+++ src/Component/spec/Doctrine/Persistence/InMemoryRepositorySpec.php
@@ -16,9 +16,9 @@ namespace spec\Sylius\Resource\Doctrine\Persistence;
 use Pagerfanta\Pagerfanta;
 use PhpSpec\ObjectBehavior;
 use spec\Sylius\Component\Resource\Fixtures\SampleBookResourceInterface;
-use Sylius\Component\Resource\Exception\UnexpectedTypeException;
 use Sylius\Resource\Doctrine\Persistence\Exception\ResourceExistsException;
 use Sylius\Resource\Doctrine\Persistence\RepositoryInterface;
+use Sylius\Resource\Exception\UnexpectedTypeException;
 use Sylius\Resource\Model\ResourceInterface;
 
 require_once dirname(__DIR__, 3) . '/legacy/spec/Fixtures/SampleBookResourceInterface.php';
